Add matching reader methods to Configuration for each option that already has a writer method.
Rename the misleading variable name @default_options to just @options, because the writer methods
(such as margin=) actually *modify* this hash, so they are not just the default values but rather
the current values.

Replace attr_accessor :default_options with a read-only to_h method for when you need to convert the
Configuration option to a Hash.

Add a Shrimp.config alias for Shrimp.configuration.