Simple change to make the Middleware thread-safe: Only make changes to instance variables on a copy of the self.
By making a copy of the middleware object, even if two web server threads started out with the same
middleware object, when they call call(), a new copy is created in each thread, and only that copy
is modified, so that the threads do not sharing any data.

This is one of the approaches mentioned at http://railscasts.com/episodes/151-rack-middleware.